Split out `mediaURLToProxyConfig`
