+++ client/lib/media/utils/index.js
@@ -8,7 +8,6 @@ const debug = debugFactory( 'calypso:media' );
  * Internal dependencies
  */
 import { reduxGetState } from 'lib/redux-bridge';
-import { getUrlParts } from 'lib/url';
 import wpcom from 'lib/wp';
 import { getEditorPostId } from 'state/ui/editor/selectors';
 
@@ -33,45 +32,7 @@ export { isItemBeingUploaded } from 'lib/media/utils/is-item-being-uploaded';
 export { isTransientPreviewable } from 'lib/media/utils/is-transient-previewable';
 export { createTransientMedia } from 'lib/media/utils/create-transient-media';
 export { validateMediaItem } from 'lib/media/utils/validate-media-item';
-
-/**
- * Given a media file URL (possibly served through photon) and site slug, returns information
- * required to correctly proxy the asset through the media proxy. Specifically, it returns
- * an object with the following keys:
- * - query: query string extracted from url
- * - filePath: path of the file on remote site, even if url is photon url
- * - isRelativeToSiteRoot: true if the file come from remote site identified by siteSlug, false otherwise
- *
- * @param {string} mediaUrl Media file URL.
- * @param {string} siteSlug Slug of the site this file belongs to.
- * @returns {object}	Dictionary
- */
-export function mediaURLToProxyConfig( mediaUrl, siteSlug ) {
-	const { pathname, search: query, protocol, hostname } = getUrlParts( mediaUrl );
-	let filePath = pathname;
-	let isRelativeToSiteRoot = true;
-
-	if ( [ 'http:', 'https:' ].indexOf( protocol ) === -1 ) {
-		isRelativeToSiteRoot = false;
-	} else if ( hostname !== siteSlug ) {
-		isRelativeToSiteRoot = false;
-		// CDN URLs like i0.wp.com/mysite.com/media.jpg should also be considered relative to mysite.com
-		if ( /^i[0-2]\.wp\.com$/.test( hostname ) ) {
-			const [ first, ...rest ] = filePath.substr( 1 ).split( '/' );
-			filePath = '/' + rest.join( '/' );
-
-			if ( first === siteSlug ) {
-				isRelativeToSiteRoot = true;
-			}
-		}
-	}
-
-	return {
-		query,
-		filePath,
-		isRelativeToSiteRoot,
-	};
-}
+export { mediaURLToProxyConfig } from 'lib/media/utils/media-url-to-proxy-config';
 
 export const getFileUploader = () => ( file, siteId ) => {
 	// Determine upload mechanism by object type
