+++ client/lib/media/utils/media-url-to-proxy-config.js
@@ -1,0 +1,43 @@
+/**
+ * Internal dependencies
+ */
+import { getUrlParts } from 'lib/url';
+
+/**
+ * Given a media file URL (possibly served through photon) and site slug, returns information
+ * required to correctly proxy the asset through the media proxy. Specifically, it returns
+ * an object with the following keys:
+ * - query: query string extracted from url
+ * - filePath: path of the file on remote site, even if url is photon url
+ * - isRelativeToSiteRoot: true if the file come from remote site identified by siteSlug, false otherwise
+ *
+ * @param {string} mediaUrl Media file URL.
+ * @param {string} siteSlug Slug of the site this file belongs to.
+ * @returns {object}	Dictionary
+ */
+export function mediaURLToProxyConfig( mediaUrl, siteSlug ) {
+	const { pathname, search: query, protocol, hostname } = getUrlParts( mediaUrl );
+	let filePath = pathname;
+	let isRelativeToSiteRoot = true;
+
+	if ( [ 'http:', 'https:' ].indexOf( protocol ) === -1 ) {
+		isRelativeToSiteRoot = false;
+	} else if ( hostname !== siteSlug ) {
+		isRelativeToSiteRoot = false;
+		// CDN URLs like i0.wp.com/mysite.com/media.jpg should also be considered relative to mysite.com
+		if ( /^i[0-2]\.wp\.com$/.test( hostname ) ) {
+			const [ first, ...rest ] = filePath.substr( 1 ).split( '/' );
+			filePath = '/' + rest.join( '/' );
+
+			if ( first === siteSlug ) {
+				isRelativeToSiteRoot = true;
+			}
+		}
+	}
+
+	return {
+		query,
+		filePath,
+		isRelativeToSiteRoot,
+	};
+}
